+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 19.5
 + Export `Falsy` utility type.
++ Export default barrier from `rvx/store`.
 
 ## 19.4
 + Support teardown hooks in `memo`.